% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{city.name}
\alias{city.name}
\title{US ZIP Cities}
\format{
A character vector of length 19108.
}
\source{
Daniel Coven's \href{http://federalgovernmentzipcodes.us/}{web site} and
the CivicSpace US ZIP Code Database written by Schuyler Erle
\href{mailto:schuyler@geocoder.us}{schuyler@geocoder.us}, 5 August 2004.
}
\usage{
city.name
}
\description{
The United States Postal Service's official names for the cities in which
ZIP codes are contained. This vector contains unique values, sorted
alphabetically; because of this, they do not line up the other vectors in the
way \link{zip.code} and \link{zip.center} do.
}
\keyword{datasets}
