% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{US Counties}
\format{
A tibble with 3,235 rows and 3 variables:
\describe{
\item{fips}{Five-digit FIPS code (state FIPS + county FIPS)}
\item{name}{County name (type suffix such as "County", "Parish", "Borough" removed)}
\item{state}{USPS state/territory abbreviation}
}
}
\source{
Census TIGER 2020 national county reference file,
\url{https://www2.census.gov/geo/docs/reference/codes2020/national_county2020.txt}
}
\usage{
counties
}
\description{
The county subdivisions of the US states and territories.
}
\keyword{datasets}
