% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county.name}
\alias{county.name}
\title{US County Names}
\format{
A character vector of length 1,932.
}
\source{
\url{https://web.archive.org/web/20240106151642/https://transition.fcc.gov/oet/info/maps/census/fips/fips.txt}
}
\usage{
county.name
}
\description{
The name of distinct US counties.
}
\keyword{datasets}
