% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_names}
\alias{county_names}
\title{US County Names}
\format{
A character vector of length 1,925.
}
\source{
Census TIGER 2020 national county reference file,
\url{https://www2.census.gov/geo/docs/reference/codes2020/national_county2020.txt}
}
\usage{
county_names
}
\description{
The name of distinct US counties.
}
\keyword{datasets}
