% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodes.R
\docType{data}
\name{fips}
\alias{fips}
\title{2018 US Census FIPS state codes}
\format{A tibble with 52 rows and 4 variables:
\describe{
\item{reg_id}{United States Census Bureau \link{divisions} ID}
\item{div_id}{United States Census Bureau \link{regions} ID}
\item{fips}{United States Census Bureau state FIPS code}
\item{state}{State name}
...
}}
\source{
\url{https://www2.census.gov/programs-surveys/popest/geographies/2018/}
}
\usage{
fips
}
\description{
FIPS state codes were numeric and two-letter alphabetic codes defined in U.S.
Federal Information Processing Standard Publication ("FIPS PUB") 5-2 to
identify U.S. states and certain other associated areas.
}
\keyword{datasets}
