% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{info}
\alias{info}
\title{US State and Territory Statistics}
\format{A tibble with 57 rows and 9 variables:
\describe{
\item{abb}{2-letter abbreviation}
\item{population}{Population estimate as of September 26, 2019}
\item{income}{Per capita income (2017)}
\item{life_exp}{Life expectancy in years (2017-18)}
\item{murder}{Murder rate per 100,000 population (2018)}
\item{high}{Percent adult population with at least a high school degree (2019)}
\item{bach}{Percent adult population with at least a bachelor's degree or greater (2019)}
\item{heat}{Mean number of degree days (temperature requires heating) per year from 1981-2010}
}}
\usage{
info
}
\description{
Updated version of the \link[datasets:state.x77]{datasets::state.x77} matrix, which provides eight
statistics from the 1970's. This version is a modern data frame format
with updated (and alternative) statistics.
}
\keyword{datasets}
