% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\docType{data}
\name{populations}
\alias{populations}
\title{2018 US State Populations}
\format{A tibble with 52 rows and 7 variables:
\describe{
\item{fips}{State FIPS code}
\item{state}{State name}
\item{population}{Resident population estimate (2018-07-01)}
\item{adult}{Percent of resident population age 18 years and over (2018-07-01)}
...
}}
\source{
\url{https://www2.census.gov/programs-surveys/popest/datasets/2010-2018/state/detail/}
}
\usage{
populations
}
\description{
SCPRC-EST2018-18+POP-RES: Estimates of the Total Resident Population and
Resident PopulationAge 18 Years and Older for the United States, States, and
Puerto Rico: July 1, 2018
}
\keyword{datasets}
