% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodes.R
\docType{data}
\name{regions}
\alias{regions}
\title{2018 US Census Regions}
\format{A tibble with 4 rows and 2 variables:
\describe{
\item{reg_id}{United States Census Bureau region ID}
\item{region}{Statistical region}
...
}}
\source{
\url{https://www2.census.gov/programs-surveys/popest/geographies/2018/}
}
\usage{
regions
}
\description{
The United States Census Bureau defines four statistical \link{regions}, with nine
\link{divisions}.
}
\keyword{datasets}
