% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_info.R
\docType{data}
\name{state.abb}
\alias{state.abb}
\title{US state and territory abbreviations}
\format{A character vector of length 57.}
\source{
\url{https://www2.census.gov/geo/docs/reference/state.txt}
}
\usage{
state.abb
}
\description{
Character vector of 2-letter abbreviations for the state and territory names.
}
\keyword{datasets}
