% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.area}
\alias{state.area}
\title{US State Areas}
\format{A numeric vector of length 57.}
\source{
\url{https://tigerweb.geo.census.gov/tigerwebmain/Files/acs19/tigerweb_acs19_state_us.html}
}
\usage{
state.area
}
\description{
The area in square miles (converted from square meters) of the US states and territories.
}
\keyword{datasets}
