% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_info.R
\docType{data}
\name{state.center}
\alias{state.center}
\title{US state geographic centers}
\format{A character vector of length 51.}
\source{
\url{https://www2.census.gov/programs-surveys/popest/geographies/2018/}
}
\usage{
state.center
}
\description{
List with components named x and y giving the approximate geographic center
of each state in negative longitude and latitude.
}
\keyword{datasets}
