% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_info.R
\docType{data}
\name{state.division}
\alias{state.division}
\title{US state Census divisions}
\format{A character vector of length 51.}
\source{
\url{https://www2.census.gov/programs-surveys/popest/geographies/2018/}
}
\usage{
state.division
}
\description{
Factor giving state divisions (New England, Middle Atlantic, South Atlantic,
East South Central, West South Central, East North Central, West North
Central, Mountain, and Pacific).
}
\keyword{datasets}
