% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.name}
\alias{state.name}
\title{US State Names}
\format{
A character vector of length 52.
}
\source{
\url{https://www2.census.gov/geo/docs/reference/state.txt}
}
\usage{
state.name
}
\description{
The full names for the US states.
}
\keyword{datasets}
