% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_info.R
\docType{data}
\name{state.region}
\alias{state.region}
\title{US state Census regions}
\format{A character vector of length 51.}
\source{
\url{https://www2.census.gov/programs-surveys/popest/geographies/2018/}
}
\usage{
state.region
}
\description{
Factor giving the region (Northeast, South, North Central, West) that each
state (and the District of Columbia) belongs to.
}
\keyword{datasets}
