% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.x19}
\alias{state.x19}
\title{US State and Territory Statistics}
\format{A tibble with 52 rows and 9 variables:
\describe{
\item{abb}{2-letter abbreviation}
\item{population}{Population estimate as of September 26, 2019}
\item{income}{Per capita income (2017)}
\item{life_exp}{Life expectancy in years (2017-18)}
\item{murder}{Murder rate per 100,000 population (2018)}
\item{high}{Percent of population with at least a high school degree (2019)}
\item{bach}{Percent of population with at least a bachelor's degree (2019)}
\item{heat}{Mean number of "degree days" per year from 1981-2010}
}}
\usage{
state.x19
}
\description{
A matrix version of the \link{facts} tibble, used to more closely align with the
\link[datasets:state.x77]{datasets::state.x77} matrix included with R.
}
\keyword{datasets}
