% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_abbs}
\alias{state_abbs}
\title{US State Abbreviations}
\format{
A character vector of length 52.
}
\source{
\url{https://www2.census.gov/geo/docs/reference/state.txt}
}
\usage{
state_abbs
}
\description{
The 2-letter USPS abbreviations for the 50 states, District of Columbia,
and Puerto Rico. Parallel to \link{state_names}.
}
\keyword{datasets}
