% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_centers}
\alias{state_centers}
\title{US State Geographic Centers}
\format{
A list of length two, each element a numeric vector of length 52.
\describe{
\item{x}{Centroid longitudinal coordinate}
\item{y}{Centroid latitudinal coordinate}
}
}
\source{
TIGER/Web REST API (State_County layer)
}
\usage{
state_centers
}
\description{
A list with components named \code{x} and \code{y} giving the approximate geographic
centroid of each state in longitude and latitude. Parallel to \link{state_names}.
}
\keyword{datasets}
