% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-convert.R
\name{state_convert}
\alias{state_convert}
\title{Convert state identifiers}
\usage{
state_convert(x, to = c("abb", "name", "fips"))
}
\arguments{
\item{x}{A character vector of: state names, abbreviations, or FIPS codes.}

\item{to}{The format returned: \code{"abb"}, \code{"name"}, or \code{"fips"}. Defaults to
\code{"abb"}.}
}
\value{
A character vector of single format state identifiers.
}
\description{
Take a vector of state identifiers and convert to a common format.
}
\examples{
state_convert(c("AL", "Vermont", "06"))
state_convert(c("AL", "Vermont", "06"), to = "name")
state_convert(c("AL", "Vermont", "06"), to = "fips")
}
