% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_facts}
\alias{state_facts}
\title{US State Facts}
\format{
A tibble with 52 rows and 9 variables:
\describe{
\item{name}{Full state name}
\item{population}{Population estimate (September 26, 2019)}
\item{votes}{Votes in the Electoral College (following the 2010 Census)}
\item{admission}{The date on which the state was admitted to the union}
\item{income}{Per capita income (2018)}
\item{life_exp}{Life expectancy in years (2017-18)}
\item{murder}{Murder rate per 100,000 population (2018)}
\item{college}{Percent adult population with at least a bachelor's degree (2019)}
\item{heat}{Mean number of degree days (temperature requires heating) per year from 1981-2010}
}
}
\source{
\itemize{
\item Population: \url{https://www2.census.gov/programs-surveys/popest/datasets/2010-2018/state/detail/SCPRC-EST2018-18+POP-RES.csv}
\item Electoral College: \url{https://www.archives.gov/electoral-college/allocation}
\item Income: Census table ACSST1Y2018.S1903
\item Life Expectancy: \url{https://web.archive.org/web/20231129160338/https://usa.mortality.org/}
\item Murder: \url{https://ucr.fbi.gov/crime-in-the-u.s/2018/crime-in-the-u.s.-2018/tables/table-4/table-4.xls/output.xls}
\item Education: Census table ACSST1Y2018.S1501
\item Temperature: NOAA 1981-2010 Climate Normals
}
}
\usage{
state_facts
}
\description{
Updated version of the \link[datasets:state]{datasets::state.x77} matrix, which provided eight
statistics from the 1970's. This version is a modern tibble with updated
statistics. See also \link{states} for state identifiers and geography.
}
\keyword{datasets}
