% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_geo}
\alias{state_geo}
\title{US State Geography}
\format{
A tibble with 52 rows and 6 variables:
\describe{
\item{abb}{2-letter USPS abbreviation (join key)}
\item{region}{Census Bureau region}
\item{division}{Census Bureau division}
\item{area}{Land area in square miles}
\item{lat}{Center latitudinal coordinate}
\item{long}{Center longitudinal coordinate}
}
}
\source{
\itemize{
\item Regions and divisions: \url{https://www2.census.gov/programs-surveys/popest/geographies/2018/state-geocodes-v2018.xlsx}
\item Area and centroids: \url{https://tigerweb.geo.census.gov/tigerwebmain/Files/acs19/tigerweb_acs19_state_us.html}
}
}
\usage{
state_geo
}
\description{
Geographic and classificatory properties for the 50 states, District of
Columbia, and Puerto Rico. Keyed by \code{abb} to join with \link{state_ids}.
}
\keyword{datasets}
