% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_names}
\alias{state_names}
\title{US State Names}
\format{
A character vector of length 52.
}
\source{
\url{https://www2.census.gov/geo/docs/reference/state.txt}
}
\usage{
state_names
}
\description{
The full names for the 50 states, District of Columbia, and Puerto Rico.
Parallel to \link{state_abbs}.
}
\keyword{datasets}
