% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_regions}
\alias{state_regions}
\title{US State Census Regions}
\format{
A factor vector of length 52.
}
\source{
\url{https://www2.census.gov/programs-surveys/popest/geographies/2018/state-geocodes-v2018.xlsx}
}
\usage{
state_regions
}
\description{
The Census region to which each state belongs, one of four. Parallel to
\link{state_names}.
}
\details{
\enumerate{
\item Northeast
\item Midwest
\item South
\item West
}
}
\keyword{datasets}
