% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_info.R
\docType{data}
\name{states}
\alias{states}
\title{2018 US state informations}
\format{A tibble with 57 rows and 6 variables:
\describe{
\item{fips}{United States Census Bureau state FIPS code}
\item{name}{State name}
\item{abb}{United States Postal Service two-letter abbreviation}
\item{ansi}{American National Standards Institute 8-digit code}
\item{region}{United States Census Bureau \link{state_divisions}}
\item{division}{United States Census Bureau \link{state_regions}}
...
}}
\source{
\url{https://www2.census.gov/geo/docs/reference/state.txt}
}
\usage{
states
}
\description{
Various official identifying codes for the fifty US states as well as the
District of Columbia and Puerto Rico, and other territories.
}
\keyword{datasets}
