% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{territory_areas}
\alias{territory_areas}
\title{US Territory Areas}
\format{
A numeric vector of length 5.
}
\source{
TIGER/Web REST API (State_County layer)
}
\usage{
territory_areas
}
\description{
The area in square miles of the US territories (AS, GU, MP, UM, VI).
}
\keyword{datasets}
