% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{territory_centers}
\alias{territory_centers}
\title{US Territory Geographic Centers}
\format{
A list of length two, each element a numeric vector of length 5.
\describe{
\item{x}{Center longitudinal coordinate}
\item{y}{Center latitudinal coordinate}
}
}
\source{
TIGER/Web REST API (State_County layer)
}
\usage{
territory_centers
}
\description{
A list with components named \code{x} and \code{y} giving the approximate geographic
center of each territory in longitude and latitude.
}
\keyword{datasets}
