% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zip.code}
\alias{zip.code}
\title{US ZIP Codes}
\format{
A character vector of length 44336.
}
\source{
Daniel Coven's \href{http://federalgovernmentzipcodes.us/}{web site} and
the CivicSpace US ZIP Code Database written by Schuyler Erle
\href{mailto:schuyler@geocoder.us}{schuyler@geocoder.us}, 5 August 2004.
}
\usage{
zip.code
}
\description{
The United States Postal Service's 5-digit codes used to identify a
particular postal delivery area.
}
\keyword{datasets}
