% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zip_centers}
\alias{zip_centers}
\title{US ZIP Centers}
\format{
A list of length two, each element a numeric vector of length 44336.
\describe{
\item{x}{Center longitudinal coordinate}
\item{y}{Center latitudinal coordinate}
}
}
\source{
Daniel Coven's \href{https://federalgovernmentzipcodes.us/}{web site} and
the CivicSpace US ZIP Code Database written by Schuyler Erle
\href{mailto:schuyler@geocoder.us}{schuyler@geocoder.us}, 5 August 2004.
}
\usage{
zip_centers
}
\description{
A list with components named \code{x} and \code{y} giving the approximate geographic
center of each ZIP code in negative longitude and latitude.
}
\keyword{datasets}
