% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{US Counties}
\format{A tibble with 3,232 rows and 3 variables:
\describe{
\item{fips}{Federal Information Processing Standard Publication 5-2 code}
\item{name}{Census county names}
\item{state}{USPS offical state, territory abbreviation code}
}}
\source{
Kyle Walker (2020). tidycensus: Load US Census Boundary and Attribute
Data as 'tidyverse' and 'sf'-Ready Data Frames. R package version 0.9.6.
https://CRAN.R-project.org/package=tidycensus
}
\usage{
counties
}
\description{
The county subdivisions of the US states and territories.
}
\keyword{datasets}
