% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{US Counties}
\format{
A tibble with 3,232 rows and 3 variables:
\describe{
\item{fips}{Federal Information Processing Standard Publication 5-2 code}
\item{name}{Census county names}
\item{state}{USPS official state, territory abbreviation code}
}
}
\source{
\url{https://web.archive.org/web/20240106151642/https://transition.fcc.gov/oet/info/maps/census/fips/fips.txt}
}
\usage{
counties
}
\description{
The county subdivisions of the US states and territories.
}
\keyword{datasets}
