% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county.name}
\alias{county.name}
\title{US County Names}
\format{A character vector of length 19108.}
\source{
Kyle Walker (2020). tidycensus: Load US Census Boundary and Attribute
Data as 'tidyverse' and 'sf'-Ready Data Frames. R package version 0.9.6.
https://CRAN.R-project.org/package=tidycensus
}
\usage{
county.name
}
\description{
The name of distinct US counties.
}
\keyword{datasets}
