% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county.name}
\alias{county.name}
\title{US County Names}
\format{A character vector of length 19108.}
\source{
\url{https://www.nrcs.usda.gov/wps/portal/nrcs/detail/national/home/?cid=nrcs143_013697}
}
\usage{
county.name
}
\description{
The name of distinct US counties.
}
\keyword{datasets}
