% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.abb}
\alias{state.abb}
\title{US State Abbreviations}
\format{
A character vector of length 52.
}
\source{
\url{https://www2.census.gov/geo/docs/reference/state.txt}
}
\usage{
state.abb
}
\description{
The 2-letter abbreviations for the US state names.
}
\keyword{datasets}
