% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.center}
\alias{state.center}
\title{US State Centers}
\format{A list of length two, each element a numeric vector of length 52.
\describe{
\item{x}{Center longitudinal coordinate}
\item{y}{Center latitudinal coordinate}
}}
\source{
\url{https://tigerweb.geo.census.gov/tigerwebmain/Files/acs19/tigerweb_acs19_state_us.html}
}
\usage{
state.center
}
\description{
A list with components named \code{x} and \code{y} giving the approximate geographic
center of each state in negative longitude and latitude.
}
\keyword{datasets}
