% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.division}
\alias{state.division}
\title{US State Divisions}
\format{
A factor vector of length 52.
}
\source{
\url{https://www2.census.gov/programs-surveys/popest/geographies/2018/state-geocodes-v2018.xlsx}
}
\usage{
state.division
}
\description{
The Census division to which each state belongs, one of nine:
\enumerate{
\item New England
\item Middle Atlantic
\item East North Central
\item West North Central
\item South Atlantic
\item East South Central
\item West South Central
\item Mountain
\item Pacific
}
}
\keyword{datasets}
