% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_info.R
\docType{data}
\name{state.name}
\alias{state.name}
\title{US state and territory names}
\format{A character vector of length 57.}
\source{
\url{https://www2.census.gov/geo/docs/reference/state.txt}
}
\usage{
state.name
}
\description{
Character vector giving the full state and territory names.
}
\keyword{datasets}
