% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.name}
\alias{state.name}
\title{US State Names}
\format{
A numeric vector of length 52.
}
\source{
\url{https://tigerweb.geo.census.gov/tigerwebmain/Files/acs19/tigerweb_acs19_state_us.html}
}
\usage{
state.name
}
\description{
The full names for the US states.
}
\keyword{datasets}
