% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.region}
\alias{state.region}
\title{US State Regions}
\format{A factor vector of length 52.}
\source{
\url{https://www2.census.gov/programs-surveys/popest/geographies/2018/state-geocodes-v2018.xlsx}
}
\usage{
state.region
}
\description{
The Census region to which each state belongs, one of four:
\enumerate{
\item Northeast
\item Midwest
\item South
\item West
}
}
\keyword{datasets}
