% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_info.R
\docType{data}
\name{state_abbs}
\alias{state_abbs}
\title{2018 US state abbreviations}
\format{A tibble with 57 rows and 2 variables:
\describe{
\item{fips}{United States Census Bureau state FIPS code}
\item{abb}{United States Postal Service two-letter abbreviation}
...
}}
\source{
\url{https://www2.census.gov/geo/docs/reference/state.txt}
}
\usage{
state_abbs
}
\description{
The USPS two-letter abbreviations for each of the US states and territories.
}
\keyword{datasets}
