% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_areas}
\alias{state_areas}
\title{US State Land Areas}
\format{
A numeric vector of length 51.
}
\source{
TIGER/Web REST API (State_County layer)
}
\usage{
state_areas
}
\description{
Land area in square miles for the 50 states and District of Columbia.
Parallel to \link{state_names}.
}
\keyword{datasets}
