% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_capitals}
\alias{state_capitals}
\title{US State Capitals}
\format{
A tibble with 51 rows and 5 variables:
\describe{
\item{abb}{2-letter USPS abbreviation (join key)}
\item{capital}{Capital city name}
\item{lat}{Latitudinal coordinate of the capital}
\item{long}{Longitudinal coordinate of the capital}
\item{population}{Capital city population (2020 Decennial Census,
city proper)}
}
}
\source{
\url{https://www.census.gov/quickfacts/}
}
\usage{
state_capitals
}
\description{
Capital cities for the 50 states and District of Columbia,
with coordinates and 2020 Census population.
}
\keyword{datasets}
