% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_info.R
\docType{data}
\name{state_centers}
\alias{state_centers}
\title{2018 US state centers}
\format{A tibble with 56 rows and 3 variables:
\describe{
\item{fips}{United States Census Bureau state FIPS code}
\item{x}{Longitude}
\item{y}{Latitude}
...
}}
\source{
\url{https://en.wikipedia.org/wiki/List_of_geographic_centers_of_the_United_States}
}
\usage{
state_centers
}
\description{
Updated list of USGS state geographic centroids.
}
\keyword{datasets}
