% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-convert.R
\name{state_convert}
\alias{state_convert}
\title{Convert state identifiers}
\usage{
state_convert(x, to = NULL)
}
\arguments{
\item{x}{A character vector of: state names, abbreviations, or FIPS codes.}

\item{to}{The format returned: "abb", "name" or "fips".}
}
\value{
A character vector of single format state identifiers.
}
\description{
Take a vector of state identifiers and convert to a common format.
}
\examples{
state_convert(c("AL", "Vermont", "06"))
}
