% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_info.R
\docType{data}
\name{state_divisions}
\alias{state_divisions}
\title{2018 US Census Bureau divisions}
\format{A tibble with 4 rows and 2 variables:
\describe{
\item{div_id}{United States Census Bureau division ID}
\item{region}{Statistical division}
...
}}
\source{
\url{https://www2.census.gov/programs-surveys/popest/geographies/2018/}
}
\usage{
state_divisions
}
\description{
The United States Census Bureau defines four statistical \link{state_regions},
with nine \link{state_divisions}.
}
\keyword{datasets}
