% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_facts}
\alias{state_facts}
\title{US State Facts}
\format{
A tibble with 51 rows and 9 variables:
\describe{
\item{name}{Full state name}
\item{population}{Resident population (2020 Decennial Census, April 1, 2020)}
\item{electors}{Votes in the Electoral College (2020 Census reapportionment, applies 2022–2032)}
\item{admission}{The date on which the state was admitted to the union}
\item{income}{Per capita income in dollars (2022 ACS 1-year)}
\item{life_exp}{Life expectancy at birth in years, both sexes (2021 NCHS)}
\item{murder}{Homicide rate per 100,000 population (2022 FBI NIBRS)}
\item{college}{Proportion of population 25+ with a bachelor's degree or higher (2022 ACS 1-year)}
\item{frost}{Mean number of days per year with minimum temperature below freezing (1991-2020 NCEI Climate Normals)}
}
}
\source{
\itemize{
\item Population: 2020 Decennial Census PL 94-171 file, variable \code{P1_001N} via tidycensus
\item Electoral College: 2020 Census reapportionment (NARA \url{https://www.archives.gov/electoral-college/allocation})
\item Income: 2022 ACS 1-year, variable \code{B19301_001} (per capita income) via tidycensus
\item Life Expectancy: NCHS 2021 state life tables via \url{https://data.cdc.gov/api/views/it4f-frdc/rows.csv}
\item Murder: FBI Crime Data Explorer API (2022 NIBRS)
\item Education: 2022 ACS 1-year Subject Table S1501, variable \code{S1501_C02_015} via tidycensus
\item Frost: NCEI 1991-2020 Climate Normals, variable \code{ANN-TMIN-AVGNDS-LSTH032}, \url{https://www.ncei.noaa.gov/data/normals-annualseasonal/1991-2020/}
}
}
\usage{
state_facts
}
\description{
Updated version of the \link[datasets:state]{datasets::state.x77} matrix, which provided eight
statistics from the 1970s. This version is a modern tibble with updated
statistics.
}
\details{
See also \link{state_ids} for state identifiers and \link{state_geo} for geography.
}
\keyword{datasets}
