% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_geo}
\alias{state_geo}
\title{US State Geography}
\format{
A tibble with 52 rows and 10 variables:
\describe{
\item{abb}{2-letter USPS abbreviation (join key)}
\item{region}{Census Bureau region}
\item{division}{Census Bureau division}
\item{area_land}{Land area in square miles}
\item{area_water}{Water area in square miles}
\item{lat}{Centroid latitudinal coordinate}
\item{long}{Centroid longitudinal coordinate}
\item{contiguous}{\code{TRUE} for the 48 contiguous states and DC;
\code{FALSE} for Alaska, Hawaii, and Puerto Rico}
\item{landlocked}{\code{TRUE} for states with no coastline on an ocean,
gulf, or Great Lake (21 states including DC)}
\item{peak_elev}{Elevation of the state high point in feet}
}
}
\source{
\itemize{
\item Regions and divisions: \url{https://www2.census.gov/programs-surveys/popest/geographies/2018/state-geocodes-v2018.xlsx}
\item Area and centroids: TIGER/Web REST API (State_County layer)
\item Peak elevations: USGS state high point records
}
}
\usage{
state_geo
}
\description{
Geographic and classificatory properties for the 50 states, District of
Columbia, and Puerto Rico. Keyed by \code{abb} to join with \link{state_ids}.
Puerto Rico has \code{NA} for \code{region} and \code{division} as it is
not assigned to a Census region.
}
\keyword{datasets}
