% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_info.R
\docType{data}
\name{state_names}
\alias{state_names}
\title{2018 US state names}
\format{A tibble with 57 rows and 2 variables:
\describe{
\item{fips}{United States Census Bureau state FIPS code}
\item{name}{Full legal name}
...
}}
\source{
\url{https://www2.census.gov/geo/docs/reference/state.txt}
}
\usage{
state_names
}
\description{
The full name for each of the US states and territories.
}
\keyword{datasets}
