% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_info.R
\docType{data}
\name{state_regions}
\alias{state_regions}
\title{2018 US Census Bureau regions}
\format{A tibble with 4 rows and 2 variables:
\describe{
\item{reg_id}{United States Census Bureau region ID}
\item{region}{Statistical region}
...
}}
\source{
\url{https://www2.census.gov/programs-surveys/popest/geographies/2018/}
}
\usage{
state_regions
}
\description{
The United States Census Bureau defines four statistical \link{state_regions},
with nine \link{state_divisions}.
}
\keyword{datasets}
