% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{US State and Territories}
\format{
A tibble with 52 rows and 8 variables:
\describe{
\item{abb}{2-letter abbreviation}
\item{name}{Full legal name}
\item{fips}{Federal Information Processing Standard Publication 5-2 code}
\item{region}{Census Bureau region}
\item{division}{Census Bureau division}
\item{area}{Area in square miles}
\item{lat}{Center latitudinal coordinate}
\item{long}{Center longitudinal coordinate}
}
}
\usage{
states
}
\description{
The 50 states, District of Columbia, and Puerto Rico.
}
\keyword{datasets}
