% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{territory.abb}
\alias{territory.abb}
\title{US Territory Abbreviations}
\format{
A character vector of length 52.
}
\source{
\url{https://www2.census.gov/geo/docs/reference/state.txt}
}
\usage{
territory.abb
}
\description{
The 2-letter abbreviations for the US territory names.
}
\keyword{datasets}
