% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{territory.area}
\alias{territory.area}
\title{US State Areas}
\format{
A numeric vector of length 52.
}
\source{
\url{https://tigerweb.geo.census.gov/tigerwebmain/Files/acs19/tigerweb_acs19_state_us.html}
}
\usage{
territory.area
}
\description{
The area in square miles of the US territories.
}
\keyword{datasets}
