% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{territory}
\alias{territory}
\title{US Territories}
\format{
A tibble with 7 rows and 6 variables:
\describe{
\item{abb}{2-letter abbreviation}
\item{name}{Full legal name}
\item{fips}{Federal Information Processing Standard Publication 5-2 code}
\item{area}{Area in square miles}
\item{lat}{Center latitudinal coordinate}
\item{long}{Center longitudinal coordinate}
}
}
\usage{
territory
}
\description{
The 6 non-state territories and federal district.
}
\keyword{datasets}
