% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{territory_abbs}
\alias{territory_abbs}
\title{US Territory Abbreviations}
\format{
A character vector of length 6.
}
\source{
\url{https://www2.census.gov/geo/docs/reference/state.txt}
}
\usage{
territory_abbs
}
\description{
The 2-letter abbreviations for the US territories (PR, AS, GU, MP, UM, VI).
}
\keyword{datasets}
