% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{territory_names}
\alias{territory_names}
\title{US Territory Names}
\format{
A character vector of length 6.
}
\source{
\url{https://www2.census.gov/geo/docs/reference/state.txt}
}
\usage{
territory_names
}
\description{
The full names for the US territories (PR, AS, GU, MP, UM, VI).
}
\keyword{datasets}
